+++ navigator-example/src/main/java/com/zhuinden/navigatorexample/SecondView.java
@@ -1,0 +1,70 @@
+package com.zhuinden.navigatorexample;
+
+import android.annotation.TargetApi;
+import android.content.Context;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.util.AttributeSet;
+import android.util.Log;
+import android.widget.LinearLayout;
+
+import com.zhuinden.simplestack.Bundleable;
+import com.zhuinden.statebundle.StateBundle;
+
+import butterknife.ButterKnife;
+
+/**
+ * Created by Zhuinden on 2017.03.10..
+ */
+
+public class SecondView
+        extends LinearLayout
+        implements Bundleable {
+    public SecondView(Context context) {
+        super(context);
+        init(context);
+    }
+
+    public SecondView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+        init(context);
+    }
+
+    public SecondView(Context context, AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+        init(context);
+    }
+
+    @TargetApi(21)
+    public SecondView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
+        init(context);
+    }
+
+    private void init(Context context) {
+        if(!isInEditMode()) {
+            // ...
+        }
+    }
+
+    @Override
+    protected void onFinishInflate() {
+        super.onFinishInflate();
+        ButterKnife.bind(this);
+    }
+
+    @NonNull
+    @Override
+    public StateBundle toBundle() {
+        StateBundle bundle = new StateBundle();
+        bundle.putString("HELLO", "WORLD");
+        return bundle;
+    }
+
+    @Override
+    public void fromBundle(@Nullable StateBundle bundle) {
+        if(bundle != null) {
+            Log.i("SECOND", bundle.getString("HELLO"));
+        }
+    }
+}